updae valida opcao

+++ LojinhaVirtual/src/marketplace/ItemCarrinho.java
@@ -2,18 +2,18 @@ package marketplace;
 
 public class ItemCarrinho {
 	
-	private Long  qntd;
+	private Integer  qntd;
 	private Double precoTotalItem;
 	private Integer codProd;
 	
-	public ItemCarrinho(Long qntd, Double precoTotalItem, Integer codProd) {
+	public ItemCarrinho(Integer qntd, Double precoTotalItem, Integer codProd) {
 		this.qntd = qntd;
 		this.precoTotalItem = precoTotalItem;
 		this.codProd = codProd;
 	}
 
-	public Long getQntd() { return qntd; }
-	public void setQntd(Long qntd) { this.qntd = qntd; }
+	public Integer getQntd() { return qntd; }
+	public void setQntd(Integer qntd) { this.qntd = qntd; }
 
 	public Double getPrecoTotalItem() { return precoTotalItem; }
 	public void setPrecoTotal(Double precoTotalItem) { this.precoTotalItem = precoTotalItem; }
